Add filename to graph
The filename is extracted and added to the graph.
This will allow the template to use this filename in the display,
linking each graph to the file that it originated from.

ntpath is used here because of
https://stackoverflow.com/q/8384737/281089

Note that the whole filename is used, including extensions.
I have deliberately not tried to remove the extension as that will
be additional (albeit minor) work that is not strictly required.
Simply removing the extension from the filename should be sufficient.
However, if this is not adequate, this removal can be added in.